moved aggregate root test case to Aggregates folder
