+++ domain/tests/Aggregates/AggregateRootTestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Domain\Tests;
+namespace Domain\Tests\Aggregates;
 
 use App\Services\ObjectHydration\PayloadSerializerFactory;
 use EventSauce\EventSourcing\Serialization\PayloadSerializer;
@@ -10,7 +10,7 @@ abstract class AggregateRootTestCase extends BaseAggregateRootTestCase
 {
     protected function payloadSerializer(): PayloadSerializer
     {
-        $config = require __DIR__ . '/../../config/eventsourcing.php';
+        $config = require __DIR__ . '/../../../config/eventsourcing.php';
 
         return PayloadSerializerFactory::make($config['hydrator_class_map']);
     }
